+++ rich/color.py
@@ -249,7 +249,7 @@ class ColorParseError(Exception):
 RE_COLOR = re.compile(
     r"""^
 \#([0-9a-f]{6})$|
-([0-9]{1,3})$|
+color\(([0-9]{1,3})\)$|
 rgb\(([\d\s,]+)\)$
 """,
     re.VERBOSE,
@@ -338,16 +338,12 @@ class Color(NamedTuple):
         if color == "default":
             return cls(color, type=ColorType.DEFAULT)
 
-        named_color_number = ANSI_COLOR_NAMES.get(color)
-        if named_color_number is not None:
+        color_number = ANSI_COLOR_NAMES.get(color)
+        if color_number is not None:
             return cls(
                 color,
-                type=(
-                    ColorType.STANDARD
-                    if named_color_number < 16
-                    else ColorType.EIGHT_BIT
-                ),
-                number=named_color_number,
+                type=(ColorType.STANDARD if color_number < 16 else ColorType.EIGHT_BIT),
+                number=color_number,
             )
 
         color_match = RE_COLOR.match(color)
@@ -355,21 +351,21 @@ class Color(NamedTuple):
             raise ColorParseError(f"{color!r} is not a valid color")
 
         color_24, color_8, color_rgb = color_match.groups()
-        if color_8:
-            number = int(color_8)
-            if number > 255:
-                raise ColorParseError(f"8bit colors must be <= 255 in {color!r}")
-            return cls(
-                color,
-                ColorType.STANDARD if number < 16 else ColorType.EIGHT_BIT,
-                number=number,
-            )
-
-        elif color_24:
+        if color_24:
             triplet = ColorTriplet(
                 int(color_24[0:2], 16), int(color_24[2:4], 16), int(color_24[4:6], 16)
             )
             return cls(color, ColorType.TRUECOLOR, triplet=triplet)
+
+        elif color_8:
+            number = int(color_8)
+            if number > 255:
+                raise ColorParseError(f"color number must be <= 255 in {color!r}")
+            return cls(
+                color,
+                type=(ColorType.STANDARD if number < 16 else ColorType.EIGHT_BIT),
+                number=number,
+            )
 
         else:  #  color_rgb:
             components = color_rgb.split(",")