+++ tests/test_markup.py
@@ -1,15 +1,33 @@
 import pytest
 
-from rich.markup import escape, MarkupError, _parse, render, Tag
+from rich.markup import escape, MarkupError, _parse, render, Tag, RE_TAGS
 from rich.text import Span
 
 
+def test_re_no_match():
+    assert RE_TAGS.match("[True]") == None
+    assert RE_TAGS.match("[False]") == None
+    assert RE_TAGS.match("[None]") == None
+    assert RE_TAGS.match("[1]") == None
+    assert RE_TAGS.match("[2]") == None
+    assert RE_TAGS.match("[]") == None
+
+
+def test_re_match():
+    assert RE_TAGS.match("[true]")
+    assert RE_TAGS.match("[false]")
+    assert RE_TAGS.match("[none]")
+    assert RE_TAGS.match("[color(1)]")
+    assert RE_TAGS.match("[#ff00ff]")
+    assert RE_TAGS.match("[/]")
+
+
 def test_escape():
-    assert escape("foo[bar]") == "foo[[bar]]"
+    assert escape("foo[bar]") == "foo\[bar]"
 
 
 def test_parse():
-    result = list(_parse("[foo]hello[/foo][bar]world[/][[escaped]]"))
+    result = list(_parse("[foo]hello[/foo][bar]world[/]\[escaped]"))
     expected = [
         (0, None, Tag(name="foo", parameters=None)),
         (10, "hello", None),
@@ -18,8 +36,7 @@ def test_parse():
         (26, "world", None),
         (26, None, Tag(name="/", parameters=None)),
         (29, "[", None),
-        (38, "escaped", None),
-        (38, "]", None),
+        (31, "escaped]", None),
     ]
     assert result == expected
 
@@ -38,6 +55,12 @@ def test_render():
     result = render("[bold]FOO[/bold]")
     assert str(result) == "FOO"
     assert result.spans == [Span(0, 3, "bold")]
+
+
+def test_render_not_tags():
+    result = render('[[1], [1,2,3,4], ["hello"], [None], [False], [True]] []')
+    assert str(result) == '[[1], [1,2,3,4], ["hello"], [None], [False], [True]] []'
+    assert result.spans == []
 
 
 def test_render_link():